Major refactor

+++ SilkySouls/Memory/GameIds.cs
@@ -4,16 +4,16 @@
     {
         public static class EventFlags
         {
-            public static readonly ulong[] UnlockKalameet = { 0xAB0F60, 0x71D };
-            public const ulong GargBell = 0xA8028C;
-            public const ulong QuelaagBell = 0xADF408;
-            public const ulong Sens = 0xAF7AA8;
-            public const ulong PlaceLordVessel = 0xB40E24;
-            public const ulong DukesAfterLordVessel = 0xB28776;
-            public const ulong NewLondoWater = 0xB100E5;
-            public const ulong LaurentiusToFirelink = 0x4E3;
-            public static readonly ulong[] GriggsToFirelink = {0x458, 0xA82996};
-            public static readonly ulong[] LoganToFirelink = {0x443, 0xAF7C32};
+            public static readonly int[] UnlockKalameet = { 0xAB0F60, 0x71D };
+            public const int GargBell = 0xA8028C;
+            public const int QuelaagBell = 0xADF408;
+            public const int Sens = 0xAF7AA8;
+            public const int PlaceLordVessel = 0xB40E24;
+            public const int DukesAfterLordVessel = 0xB28776;
+            public const int NewLondoWater = 0xB100E5;
+            public const int LaurentiusToFirelink = 0x4E3;
+            public static readonly int[] GriggsToFirelink = {0x458, 0xA82996};
+            public static readonly int[] LoganToFirelink = {0x443, 0xAF7C32};
         }
 
         public static class ShopParams
